Hotfix- reference error from directory update
So, it turns out some core files got renamed, i didn't expect that to happen , so i just blindly imported them, some sneek peak's at the new single replace is in here too, it should be out sometime today or tomorrow, as will reinstating decking as an armor choice. unfortunately that has to be done manually.