Single paint/replace is live!
the single paint/replace tool is live, it will allow you to replace any block, with any other block, even if the placement is invalid. feel free to replace a solid block of 1m cubes with 4m beams, or place the unplaceable nuclear parts. single replace doesn't replace one component, but all blocks with the same id, and can be restricted with paint colors, and the area restrictions. furthermore you can replace the same block with itself, and just have it paint.
The next updates are likely to be more internal, expect mild increases in processing speeds, and the hiding/disabling of buttons until they are valid, such as hiding the editing tools until a blueprint is loaded and, if i'm thinking of overachieving; loading screens/alerts.

After my blunder before i'll be more careful when updating the dictionary, though the importer tool now has many more failsafes, so hopefully it wont be a problem.