Behind the scenes update
the editor has had some changes, i've also manually edited the dictionary, decking should be convertible in addition to paintable, and app armar should be too, but i've done less testing with that. also the page hides until a BP is supplied, uncertain if this will be a permanent change. besides that, ive sone some fiddling with converters, hopefully it's faster. i added an alert to bp changes, before immediately ripping it out, vanilla alerts are annoying, so ill probably just have the DL button flash on completion or something. decisions for another day i suppose